checkpoint

+++ const.go
@@ -34,16 +34,16 @@ const (
 	KeyError        = 0x31
 
 	// https://github.com/fl00r/go-tarantool-1.6/issues/2
-	IterEq            = uint32(0) // key == x ASC order
-	IterReq           = uint32(1) // key == x DESC order
-	IterAll           = uint32(2) // all tuples
-	IterLt            = uint32(3) // key < x
-	IterLe            = uint32(4) // key <= x
-	IterGe            = uint32(5) // key > x
-	IterGt            = uint32(6) // key >= x
-	IterBitsAllSet    = uint32(7) // all bits from x are set in key
-	IterBitsAnySet    = uint32(8) // at least one x's bit is set
-	IterBitsAllNotSet = uint32(9) // all bits are not set
+	IterEq            = uint8(0) // key == x ASC order
+	IterReq           = uint8(1) // key == x DESC order
+	IterAll           = uint8(2) // all tuples
+	IterLt            = uint8(3) // key < x
+	IterLe            = uint8(4) // key <= x
+	IterGe            = uint8(5) // key > x
+	IterGt            = uint8(6) // key >= x
+	IterBitsAllSet    = uint8(7) // all bits from x are set in key
+	IterBitsAnySet    = uint8(8) // at least one x's bit is set
+	IterBitsAllNotSet = uint8(9) // all bits are not set
 
 	OkCode            = uint32(0)
 	PacketLengthBytes = 5
